requirement.txt
anyio==3.6.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.85.0
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
Werkzeug==2.2.2
zipp==3.8.1
